Split up giant `test_user` file
The only two files left that are not clearly associated to one
functionality are

- test_user, where the one giant „move in and back out“ test case
  lives
- test_dormant_user, where moving and moving out are both tested.

Splitting tests up into modules is a good thing towards a more
pytest-like structure, where instead of base classes and setup
routines we will have  module-based grouping and fixtures.
