+++ tests/lib/user/__init__.py
@@ -1,0 +1,7 @@
+class ExampleUserData:
+    name = u"Hans"
+    login = u"hans66"
+    email = u"hans@hans.de"
+    mac = "12:11:11:11:11:11"
+    birthdate = "1990-01-01"
+
